requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
dill==0.3.8
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gunicorn==23.0.0
huggingface-hub==0.30.1
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
multiprocess==0.70.16
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
zstd==1.5.6.7
